handle degenerate bbox
A ST_Intersects() with a bounding box / polygon that is a point or a horizontal/vertical line results in an empty results (since it is an invalid polygon). Reduce an envelope that is a point to a point and reduce an envelope that is a horizontal/vertical line to a line string.

Closes https://github.com/interactive-instruments/ldproxy/issues/709