Only seed layers with features
If a feature type has no features, the spatial extent of the collection is empty. In this case the bounding box of the tile matrix set was used as a fallback for the spatial extent, which for WebMercatorQuad results in seeding the whole world. Instead the layer / collection has to be skipped.