Fix detection of time elements in dateFormat transformation (#194)
A date should remain as a local date, if the pattern does not include time aspects. If it includes time aspects, the date is converted to the start of the day in the default timezone and then converted to UTC. The detection was faulty as it also included "M", which represents months. That is, also date formats were detected to include time aspects.