Tiles: Improve handling of configuration issues
If a tile provider from features currently includes layers that reference non-existing tiles, an exception is thrown that does not provide a hint how to fix the configuration. This change throws an error, if a layer without a feature type is associated. If the feature type has no primary geometry, the layer will now exclude all features instead of throwing an error.

The configuration should be validated during provider startup.