initialize support for Last-Modified
A feature property can be marked with 'isLastModified: true' in the provider schema to declare that the value is the timestamp when the feature has last been updated.

The value is added to FeatureStream.Result in addition to the ETag. The value is not yet set. We should first discuss how to implement this. There is a TODO in the FeatureStream class with more information.