fix handling CRUD requests with FEATURE_REF properties
When creating the INSERT statements for a column with type FEATURE_REF, the value was not quoted. As a result, inserting the value 0-20000-0-60191 resulted in the value -80191 in the database. This change quotes also FEATURE_REF values of value type STRING.

Note: No issue is required since FEATURE_REF will only be introduced in v3.5.0.