queryable dynamic feature refs
Currently, only static feature references (fixed `refType`) are handled in `FeatureRefResolver`. There are, however, cases where dynamic feature references can be queryable, too. This is the case, if the type is determined by a `type` property that is not constant, but takes the `refType` value from a column. No concat/coalesce is used in this case..