queryable dynamic feature refs (#236)
Currently, only static feature references (fixed `refType`) are handled in `FeatureRefResolver`. There are, however, cases where dynamic feature references can be queryable, too. This is the case, if the type is determined by a `type` property that is not constant, but takes the `refType` value from a column. No concat/coalesce is used in this case. However, such queryables are only useful, if the feature ids are globally unique. 

Currently, the OGC API Features Part 3 does not specify how a queryable that is a feature reference which has more variables than the local `featureId` should be interpreted. In this sense, the current approach is a temporary solution that may change in the future. I will create a separate PR to update the documentation of the Queryables module.