add checks on queryable properties (#291)
* check queryable/sortable properties

Properties for which the value is transformed should not be queryables or sortable as the results may be unexpected. However, this would be a breaking change. Until the next major version, we should issue a warning in the log. When using the wildcard (‘*’), however, I would rather suppress such properties.

However, this change is not sufficient for FEATURE_REF properties where the "id" property has a value transformation as the transformation is not included in the queryables schema. This is an open issue.

* update feature ref checks

* add feature ref id transformations to queryable